Finalise support for ad-hoc items
Finalise the already existing partial support for "ad-hoc" items,
which are items on the shopping list that are not permanently configured
for any shop.

It is now also possible to add an item both to the list and the shop
at the same time with a new "add to shop" checkbox for adding
not-yet-configured items.